+++ graphql-maven-plugin-logic/src/main/resources/templates/server_GraphQLDataFetchers.vm.java
@@ -59,19 +59,15 @@ public class GraphQLDataFetchers {
 #end
 
 #if (${dataFetcher.completableFuture})
-			DataLoader<${dataFetcher.field.type.identifier.javaTypeFullClassname}, ${dataFetcher.field.javaTypeFullClassname}> dataLoader = dataFetchingEnvironment.getDataLoader("${dataFetcher.field.type.classSimpleName}"); 
+			DataLoader<${dataFetcher.field.type.identifier.javaTypeFullClassname}, ${dataFetcher.field.type.classFullName}> dataLoader = dataFetchingEnvironment.getDataLoader("${dataFetcher.field.type.classSimpleName}"); 
 			
 			// This dataLoader may be null. Let's handle that:
 			if (dataLoader != null) { 
 				return ${dataFetchersDelegate.camelCaseName}.${dataFetcher.javaName}(dataFetchingEnvironment, dataLoader#if($dataFetcher.graphQLOriginType), source#end#foreach($argument in $dataFetcher.field.inputParameters), ${argument.javaName}#end);
 			} else {
-#if ($configuration.legacyDataLoaderCall)
 				logger.warn("No DataLoader found for key '{}'", "${dataFetcher.field.type.classSimpleName}");
 				return CompletableFuture.supplyAsync(
 						() -> ${dataFetchersDelegate.camelCaseName}.${dataFetcher.javaName}(dataFetchingEnvironment#if($dataFetcher.graphQLOriginType), source#end#foreach($argument in $dataFetcher.field.inputParameters), ${argument.javaName}#end));
-#else
-				throw new NullPointerException("The dataloader for ${dataFetcher.field.type.classSimpleName} may not be null");
-#end
 			}
 #elseif (${dataFetcher.field.fieldTypeAST.listDepth}>0)
 			#if (${dataFetcher.dataFetcherDelegate.type.requestType}=="subscription")Publisher<#if($dataFetcher.field.fieldTypeAST.mandatory==false)Optional<#end#end${dataFetcher.field.javaTypeFullClassname}#if(${dataFetcher.dataFetcherDelegate.type.requestType}=="subscription")#if($dataFetcher.field.fieldTypeAST.mandatory==false)>#end>#end ret = ${dataFetchersDelegate.camelCaseName}.${dataFetcher.javaName}(dataFetchingEnvironment#if($dataFetcher.graphQLOriginType), source#end#foreach($argument in $dataFetcher.field.inputParameters), ${argument.javaName}#end);