+++ graphql-maven-plugin-logic/src/test/java/graphql/mavenplugin_notscannedbyspring/Forum_Server_SpringConfiguration.java
@@ -35,7 +35,7 @@ public class Forum_Server_SpringConfiguration extends AbstractSpringConfiguratio
 				"com.graphql_java_generator.customscalars.GraphQLScalarTypeDate.Date", null));
 
 		configuration.schemaFilePattern = "forum.graphqls";
-		configuration.legacyDataLoaderCall = false;
+		configuration.generateDataLoaderForLists = true;
 		configuration.mode = PluginMode.server;
 		configuration.schemaPersonalizationFile = new File(mavenTestHelper.getModulePathFile(),
 				"src/test/resources/forum_personalization.json");