+++ graphql-maven-plugin/src/main/java/com/graphql_java_generator/mavenplugin/AbstractGenerateServerCodeMojo.java
@@ -93,15 +93,14 @@ public abstract class AbstractGenerateServerCodeMojo extends AbstractGenerateCod
 	 * (only for server mode) Defines how the methods in the data fetchers delegates are generated.
 	 * </P>
 	 * <P>
-	 * When legacyDataLoaderCall is true (default mode), the data loaders are used only for fields that don't return a
-	 * list. And a useless method without data loader (that is: a method that doesn't return a
-	 * {@link CompletableFuture}) for each fields that use a data loader. In other words, for fields which type is a
-	 * sub-object with an id, two methods are generated: one which returns a {@link CompletableFuture}, and one which
-	 * returns a none {@link CompletableFuture} result (and that is not used by the generated code).
+	 * When generateDataLoaderForLists is false (default mode), the data loaders are used only for fields that don't
+	 * return a list. In other words, for fields which type is a sub-object with an id, two methods are generated: one
+	 * which returns a {@link CompletableFuture}, and one which returns a none {@link CompletableFuture} result (that is
+	 * used by the generated code only if no data loader is available).
 	 * </P>
 	 * <P>
-	 * When legacyDataLoaderCall is false, there is one getter for field of this data fetcher. If the field's type is a
-	 * type with an id (whether it is a list or not), then the return type is a {@link CompletableFuture}. If the
+	 * When generateDataLoaderForLists is true, there is one getter for field of this data fetcher. If the field's type
+	 * is a type with an id (whether it is a list or not), then the return type is a {@link CompletableFuture}. If the
 	 * field's type is a type that has no id, then a method is generated with a direct return (not a
 	 * {@link CompletableFuture}). When the field's type is a scalar, then no method is generated (no need to fetch it)
 	 * </P>
@@ -109,8 +108,8 @@ public abstract class AbstractGenerateServerCodeMojo extends AbstractGenerateCod
 	 * This parameter is available since version 1.18.4
 	 * </P>
 	 */
-	@Parameter(property = "com.graphql_java_generator.mavenplugin.legacyDataLoaderCall", defaultValue = GraphQLConfiguration.DEFAULT_LEGACY_DATA_LOADER_CALL)
-	boolean legacyDataLoaderCall;
+	@Parameter(property = "com.graphql_java_generator.mavenplugin.legacyDataLoaderCall", defaultValue = GraphQLConfiguration.DEFAULT_GENERATE_DATA_LOADER_FOR_LISTS)
+	boolean generateDataLoaderForLists;
 
 	/**
 	 * <P>
@@ -206,8 +205,8 @@ public abstract class AbstractGenerateServerCodeMojo extends AbstractGenerateCod
 	}
 
 	@Override
-	public boolean isLegacyDataLoaderCall() {
-		return legacyDataLoaderCall;
+	public boolean isGenerateDataLoaderForLists() {
+		return generateDataLoaderForLists;
 	}
 
 	@Override